agents/image-gen-a2a-server/requirements.txt
requests==2.32.4
a2a-sdk==0.2.6
click==8.2.1
uvicorn==0.35.0